Only switch to alternate buffer for fancy ui
