+++ bin.js
@@ -133,7 +133,6 @@ var argv = minimist(process.argv.slice(2), {
   } else if (cmd === 'inspect') {
     require('./lib/cmd-inspect')(path.join(entry), argv)
   } else if (cmd === 'start') {
-    if (!argv.q) alternateBuffer()
     require('./lib/cmd-start')(path.join(entry), argv)
   } else {
     console.log(NOCOMMAND)
@@ -143,51 +142,4 @@ var argv = minimist(process.argv.slice(2), {
 
 function clr (text, color) {
   return process.stdout.isTTY ? ansi.format(text, color) : text
-}
-
-// Switch to an alternate terminal buffer,
-// switch back to the main terminal buffer on exit.
-function alternateBuffer () {
-  var q = Buffer.from('q')
-  var esc = Buffer.from([0x1B])
-
-  process.stdout.write('\x1b[?1049h') // Enter alternate buffer.
-  process.stdout.write('\x1b[H') // Reset screen to top.
-  process.stdout.write('\x1b[?25l') // Hide cursor
-
-  process.on('unhandledRejection', onexit)
-  process.on('uncaughtException', onexit)
-  process.on('SIGTERM', onexit)
-  process.on('SIGINT', onexit)
-  process.on('exit', onexit)
-  process.stdin.on('data', handleKey)
-
-  function handleKey (buf) {
-    if (buf.compare(q) === 0 || buf.compare(esc) === 0) {
-      onexit()
-    }
-  }
-
-  function onexit (statusCode) {
-    process.stdout.write('\x1b[?1049l') // Enter to main buffer.
-    process.stdout.write('\x1b[?25h') // Restore cursor
-
-    if (statusCode instanceof Error) {
-      console.error('A critical error occured, forcing Bankai to abort:\n')
-      console.error(clr(statusCode.stack, 'red') + '\n')
-      console.error(dedent`
-        If you think this might be a bug in Bankai, please consider helping
-        improve Bankai's stability by submitting an error to:
-
-        ${'  ' + clr('https://github.com/choojs/bankai/issues/new', 'underline')}
-
-        Please include the steps to reproduce this error, the stack trace
-        printed above, your version of Node, and your version of npm. Thanks!
-        ${clr('— Team Choo', 'italic')}
-      ` + '\n')
-      statusCode = 1
-    }
-
-    process.exit(statusCode)
-  }
 }
