+++ lib/cmd-start.js
@@ -3,7 +3,8 @@ var getPort = require('get-port')
 var isElectronProject = require('./is-electron-project')
 var http = require('./http-server')
 var bankai = require('../http')
-var ui = require('./ui')
+var createTui = require('./ui')
+var createLogUi = require('./ui-basic')
 
 module.exports = start
 
@@ -23,7 +24,8 @@ function start (entry, opts) {
       })
     })
 
-    var render = ui(handler.compiler, state)
+    var createUi = opts.simple ? createLogUi : createTui
+    var render = createUi(handler.compiler, state)
     render()
 
     getPort({port: 8080})
