+++ lib/ui-basic.js
@@ -1,6 +1,5 @@
 var nanoraf = require('nanoraf')
 var pretty = require('prettier-bytes')
-module.exports = render
 
 module.exports = createLogUI
